Google Firebase Starter Project
Do Module 9 Assignment work that includes making a JavaFX/FXML Maven project, importing over a sample application to run, and setup a Google Firebase project. With all of this set up, there are documentations that includes screenshots that show that the project and the cloud server is working. I did step 4 (write button) before step 3 (read button) so there would be data in the text area in the GUI.